minor code clean
